requirements.txt
django
django-cors-headers
whitenoise

#AI packages
transformers
scipy
torch
pillow